+++ sdk/azidentity/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 0.13.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 0.13.1 (2022-02-08)
 
 ### Features Added